✨ (appsettings.json): add new field mapping configurations for FieldMergeMap and FieldLiteralMap ♻️ (docs): update XML documentation to reflect new processor class names ♻️ (WorkItemMigrationProcessor): rename class and options to TfsWorkItemMigrationProcessor for clarity
The new field mapping configurations in `appsettings.json` provide additional flexibility for mapping fields during migration. The renaming of the `WorkItemMigrationProcessor` class and its options to `TfsWorkItemMigrationProcessor` improves clarity and consistency, making it clear that the processor is specific to TFS (Team Foundation Server). The XML documentation is updated to reflect these changes, ensuring that the documentation remains accurate and helpful.

♻️ (refactor): rename WorkItemMigrationProcessorOptions to TfsWorkItemMigrationProcessorOptions
✨ (feature): add overwrite option to InitMigrationCommand for config file handling
🔧 (config): update launchSettings.json to include new init options

Renaming `WorkItemMigrationProcessorOptions` to `TfsWorkItemMigrationProcessorOptions` improves clarity and specificity, indicating that the processor options are specific to TFS. The new `overwrite` option in `InitMigrationCommand` allows users to overwrite existing configuration files, enhancing flexibility and preventing accidental overwrites. The `launchSettings.json` update includes new initialization options, providing more configuration presets for different use cases.

♻️ (MigrationConfigCommand.cs, OptionsManager.cs): refactor configuration management and add template application support

Refactor the configuration management logic to improve readability and maintainability. Introduce a new method `AddOptionsToConfiguration` to handle adding or updating options in the configuration JSON. This change centralizes the logic for modifying the configuration, making it easier to manage and extend.

Additionally, add support for applying templates in `MigrationConfigCommand.cs`. This feature allows users to select and apply predefined templates to their configuration, enhancing the flexibility and usability of the migration tool.

+++ src/MigrationTools.Host/Commands/InitMigrationCommand.cs
@@ -1,34 +1,33 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.ApplicationInsights.DataContracts;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using MigrationTools._EngineV1.Configuration;
+using MigrationTools.Options;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 using Spectre.Console.Cli;
 
 namespace MigrationTools.Host.Commands
 {
     internal class InitMigrationCommand : AsyncCommand<InitMigrationCommandSettings>
     {
-        private readonly IEngineConfigurationBuilder _configurationBuilder;
-        private readonly ISettingsWriter _settingWriter;
         private readonly ILogger _logger;
         private readonly ITelemetryLogger Telemetery;
         private readonly IHostApplicationLifetime _appLifetime;
 
         public InitMigrationCommand(
-            IEngineConfigurationBuilder configurationBuilder,
-            ISettingsWriter settingsWriter,
             ILogger<InitMigrationCommand> logger,
             ITelemetryLogger telemetryLogger,
             IHostApplicationLifetime appLifetime)
         {
-            _configurationBuilder = configurationBuilder;
-            _settingWriter = settingsWriter;
             _logger = logger;
             Telemetery = telemetryLogger;
             _appLifetime = appLifetime;
@@ -44,44 +43,68 @@ namespace MigrationTools.Host.Commands
                 string configFile = settings.ConfigFile;
                 if (string.IsNullOrEmpty(configFile))
                 {
-                    configFile = "configuration.json";
+                    configFile = $"configuration-{settings.Options.ToString()}.json";
                 }
                 _logger.LogInformation("ConfigFile: {configFile}", configFile);
                 if (File.Exists(configFile))
                 {
-                    _logger.LogInformation("Deleting old configuration.json reference file");
-                    File.Delete(configFile);
+                    if (settings.Overwrite)
+                    {
+                        File.Delete(configFile);
+                    }
+                    else
+                    {
+                        _logger.LogCritical($"The config file {configFile} already exists, pick a new name. Or Set --overwrite");
+                        Environment.Exit(1);
+                    }
                 }
                 if (!File.Exists(configFile))
                 {
+                    var configuration = new ConfigurationBuilder()
+                    .SetBasePath(Directory.GetCurrentDirectory())
+                    .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
+                    .Build();
+                    // get source IOptions bits
+                    List<Type> allMigrationTypes = AppDomain.CurrentDomain.GetMigrationToolsTypes().ToList();
+                    var allOptions = allMigrationTypes.WithInterface<IOptions>();
+                    JObject configJson = new JObject();
+
                     _logger.LogInformation("Populating config with {Options}", settings.Options.ToString());
-                    EngineConfiguration config;
+                    List<string> optionsToInclude = null;
                     switch (settings.Options)
                     {
                         case OptionsMode.Reference:
-                            config = _configurationBuilder.BuildReference();
+
                             break;
                         case OptionsMode.Basic:
-                            config = _configurationBuilder.BuildGettingStarted();
+                             optionsToInclude = new List<string>() { "TfsWorkItemMigrationProcessorOptions", "FieldMappingToolOptions", "FieldLiteralMapOptions" };
                             break;
-
                         case OptionsMode.WorkItemTracking:
-                            config = _configurationBuilder.BuildWorkItemMigration();
+                             optionsToInclude = new List<string>() { "TfsWorkItemMigrationProcessorOptions", "FieldMappingToolOptions", "FieldLiteralMapOptions" };
                             break;
-
-                        case OptionsMode.Fullv2:
-                            config = _configurationBuilder.BuildDefault2();
-                            break;
-
-                        case OptionsMode.WorkItemTrackingv2:
-                            config = _configurationBuilder.BuildWorkItemMigration2();
-                            break;
-
                         default:
-                            config = _configurationBuilder.BuildGettingStarted();
+                            optionsToInclude = new List<string>() { "TfsWorkItemMigrationProcessorOptions", "FieldMappingToolOptions", "FieldLiteralMapOptions" };
                             break;
                     }
-                    _settingWriter.WriteSettings(config, configFile);
+
+                    if (optionsToInclude != null)
+                    {
+                        foreach (var item in optionsToInclude)
+                        {
+                            var item2 = allOptions.FirstOrDefault(x => x.Name == item);
+                            configJson = AddOptionToConfig(configuration, configJson, item2);
+                        }
+                    } else
+                    {
+                        _logger.LogWarning($"You are adding all of the Options, there may be some that cant be added and will cause an error...");
+                        foreach (var item in allOptions)
+                        {
+                            configJson = AddOptionToConfig(configuration, configJson, item);
+                        }
+                    }
+
+
+                    File.WriteAllText(configFile, configJson.ToString(Formatting.Indented));
                     _logger.LogInformation($"New {configFile} file has been created");
                 }
                 _exitCode = 0;
@@ -98,6 +121,26 @@ namespace MigrationTools.Host.Commands
                 _appLifetime.StopApplication();
             }
             return _exitCode;
+        }
+
+        private JObject AddOptionToConfig(IConfigurationRoot configuration, JObject configJson, Type item)
+        {
+            IOptions instanceOfOption = (IOptions)Activator.CreateInstance(item);
+            bool isCollection = !string.IsNullOrEmpty(instanceOfOption.ConfigurationCollectionPath);
+            var section = configuration.GetSection(instanceOfOption.ConfigurationSectionPath);
+            section.Bind(instanceOfOption);
+            try
+            {
+                configJson = Options.OptionsManager.AddOptionsToConfiguration(configJson, instanceOfOption, isCollection);
+                _logger.LogInformation("Adding Option: {item}", item.Name);
+            }
+            catch (Exception)
+            {
+
+                _logger.LogInformation("FAILED!! Adding Option: {item}", item.FullName);
+            }
+
+            return configJson;
         }
     }
 }
